Backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
requests
google-genai
langchain
langchain-google-genai